Rename const module to consts
